Separate ArrayFieldTemplate
