+++ packages/react-jsonschema-form-bootstrap/src/components/templates/index.js
@@ -1,3 +1,4 @@
+import ArrayFieldTemplate from './ArrayFieldTemplate';
 import FieldTemplate from './FieldTemplate';
 
-export default { FieldTemplate };
+export default { ArrayFieldTemplate, FieldTemplate };
