+++ playground/src/app.js
@@ -352,15 +352,11 @@ class App extends Component {
   }
 
   load = data => {
-    // Reset the ArrayFieldTemplate whenever you load new data
-    const { ArrayFieldTemplate, ObjectFieldTemplate } = data;
     // force resetting form component instance
     this.setState({ form: false }, () =>
       this.setState({
         ...data,
-        form: true,
-        ArrayFieldTemplate,
-        ObjectFieldTemplate
+        form: true
       })
     );
   };
@@ -406,8 +402,7 @@ class App extends Component {
       validate,
       theme,
       editor,
-      ArrayFieldTemplate,
-      ObjectFieldTemplate,
+      templates,
       transformErrors
     } = this.state;
 
@@ -462,8 +457,6 @@ class App extends Component {
         <div className="col-sm-5">
           {this.state.form && (
             <Form
-              ArrayFieldTemplate={ArrayFieldTemplate}
-              ObjectFieldTemplate={ObjectFieldTemplate}
               liveValidate={liveValidate}
               schema={schema}
               uiSchema={uiSchema}
@@ -473,6 +466,7 @@ class App extends Component {
                 console.log('submitted formData', formData)
               }
               fields={{ geo: GeoPosition }}
+              templates={templates}
               validate={validate}
               onBlur={(id, value) =>
                 console.log(`Touched ${id} with value ${value}`)
