+++ playground/src/samples/customArray.js
@@ -56,5 +56,5 @@ module.exports = {
     }
   },
   formData: ['react', 'jsonschema', 'form'],
-  ArrayFieldTemplate
+  templates: { ArrayFieldTemplate }
 };
