+++ playground/src/samples/customObject.js
@@ -62,5 +62,5 @@ module.exports = {
     bio: 'Roundhouse kicking asses since 1940',
     password: 'noneed'
   },
-  ObjectFieldTemplate
+  templates: { ObjectFieldTemplate }
 };
